+++ src/vm/array_wrapper.h
@@ -13,7 +13,7 @@ public:
   static string ToString(Object *obj);
 
   static Object *NewObjectArrayWrapper(VM *vm, int size);
-  static Object *NewIntArrayWrapper(VM *vm, int size,
+  static Object *NewIntArrayWrapper(VM *vm, uint64_t size,
 				    const iroha::NumericWidth &width,
 				    Annotation *an);
   static Object *Copy(VM *vm, Object *obj);
